requirements.txt
pandas
langchain
sentence-transformers
jax
jaxlib
"jax[cpu]"
google-api-python-client
openai
tabulate
ipython
psycopg2-binary
transformers
einops
accelerate
xformers
